Refactor WalletConnect
This commit implements a refactor around `WalletConnect`, with some
logical changes, but should not change any of the overall functionality.

The refactor consists of:
1. Added documentation to several functions and structs
2. Hierarchical namespacing of functions and structures to facilitate
   and indicate the context of each of them
3. Split some types into separate files, to avoid overwhelming
   information in a single file
4. Restructured WalletConnect.Request into an enum for more clear
   usage semantics

This is not an exhaustive refactor, but hopefully it makes the code
around WalletConnect easier to read and work with.
