+++ damus/Views/DMChatView.swift
@@ -131,7 +131,7 @@ struct DMChatView: View, KeyboardReadable {
             .map(\.asString)
             .joined(separator: "")
 
-        guard let dm = create_dm(content, to_pk: pubkey, tags: tags, keypair: damus_state.keypair) else {
+        guard let dm = NIP04.create_dm(content, to_pk: pubkey, tags: tags, keypair: damus_state.keypair) else {
             print("error creating dm")
             return
         }
@@ -174,46 +174,6 @@ struct DMChatView_Previews: PreviewProvider {
 
         DMChatView(damus_state: test_damus_state, dms: model)
     }
-}
-
-func encrypt_message(message: String, privkey: Privkey, to_pk: Pubkey, encoding: EncEncoding = .base64) -> String? {
-    let iv = random_bytes(count: 16).bytes
-    guard let shared_sec = get_shared_secret(privkey: privkey, pubkey: to_pk) else {
-        return nil
-    }
-    let utf8_message = Data(message.utf8).bytes
-    guard let enc_message = aes_encrypt(data: utf8_message, iv: iv, shared_sec: shared_sec) else {
-        return nil
-    }
-    
-    switch encoding {
-    case .base64:
-        return encode_dm_base64(content: enc_message.bytes, iv: iv)
-    case .bech32:
-        return encode_dm_bech32(content: enc_message.bytes, iv: iv)
-    }
-    
-}
-
-func create_encrypted_event(_ message: String, to_pk: Pubkey, tags: [[String]], keypair: FullKeypair, created_at: UInt32, kind: UInt32) -> NostrEvent? {
-    let privkey = keypair.privkey
-    
-    guard let enc_content = encrypt_message(message: message, privkey: privkey, to_pk: to_pk) else {
-        return nil
-    }
-    
-    return NostrEvent(content: enc_content, keypair: keypair.to_keypair(), kind: kind, tags: tags, createdAt: created_at)
-}
-
-func create_dm(_ message: String, to_pk: Pubkey, tags: [[String]], keypair: Keypair, created_at: UInt32? = nil) -> NostrEvent?
-{
-    let created = created_at ?? UInt32(Date().timeIntervalSince1970)
-
-    guard let keypair = keypair.to_full() else {
-        return nil
-    }
-    
-    return create_encrypted_event(message, to_pk: to_pk, tags: tags, keypair: keypair, created_at: created, kind: 4)
 }
 
 extension View {
